refactor: 💡 move navigation functionality from the World class
